+++ common/include/public.h
@@ -37,7 +37,7 @@ struct main_mngr_info
 	workstate_e work_state;
 	int user_handle;	// user client socket handle index
 	int mngr_handle;	// manager client socket handle index
-	long attend_time;
+	uint32_t attend_time;
 };
